enhanced by dto

+++ src/app/components/news-detail/news-detail.component.ts
@@ -1,48 +1,52 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NewsService } from '../../services/news.service';
+import { News } from "../../models/news.model";
 
 @Component({
-  selector: 'app-news-detail',
-  templateUrl: './news-detail.component.html',
-  styleUrls: ['./news-detail.component.scss']
+    selector: 'app-news-detail',
+    templateUrl: './news-detail.component.html',
+    styleUrls: ['./news-detail.component.scss']
 })
 export class NewsDetailComponent implements OnInit {
-  news: any;
+    news: News | null = null;
 
-  constructor(
-    private route: ActivatedRoute,
-    private router: Router,
-    private newsService: NewsService
-  ) {
-  }
+    constructor(
+        private route: ActivatedRoute,
+        private router: Router,
+        private newsService: NewsService
+    ) {
+    }
 
-  ngOnInit(): void {
-    const newsId = this.route.snapshot.paramMap.get('id');
-    if (newsId) {
-      this.newsService.getNewsById(newsId).subscribe(
-        (data) => {
-          this.news = data;
-        },
-        (error) => {
-          console.error('Error fetching news detail:', error);
+    ngOnInit(): void {
+        const newsId = this.route.snapshot.paramMap.get('id');
+        if (newsId) {
+            this.loadNews(newsId);
         }
-      );
     }
-  }
 
-  loadNews(id: string) {
-    this.newsService.getNewsById(id).subscribe(data => {
-      this.news = data;
-    });
-  }
-
-  deleteNews(): void {
-    if (confirm('Are you sure you want to delete this news item?')) {
-      this.newsService.deleteNews(this.news.id).subscribe(() => {
-        this.router.navigate(['/news']);
-      });
+    loadNews(id: string) {
+        this.newsService.getNewsById(id).subscribe(
+            (data: News) => {
+                this.news = data;
+                this.news.id = id;
+            },
+            (error) => {
+                console.error('Error fetching news detail:', error);
+            }
+        );
     }
-  }
+
+    deleteNews(): void {
+        if (this.news && this.news.id !== undefined) {
+            if (confirm('Are you sure you want to delete this news item?')) {
+                this.newsService.deleteNews(this.news.id).subscribe(() => {
+                    this.router.navigate(['/news']);
+                });
+            }
+        } else {
+            console.error('News item is not loaded or ID is undefined.');
+        }
+    }
 
 }
